requirements.txt
matplotlib==3.7.2
model==0.6.0
numpy==1.24.3
pandas==2.0.3
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.3.2
torch==1.11.0
torch_geometric==2.0.4
transformers==4.46.3
utils==1.0.2
wandb==0.21.3
